Robby Death + My Death
The animation is slowly draining away my soul and CPU power. It haunts my very being and would not interpolate correctly, for that I have only myself to blame. On this dark, dark night of sleeplessness and restlessness, I shall push this commit in hopes of coming back to it with a new hope never seen before.